log: Add `--with-feature-counts=...` option
Adds feature count [estimate] to commits in JSON output
